+++ terraform/environments/demo/cloudwatch_logs/versions.tf
@@ -2,6 +2,7 @@ terraform {
   required_version        = "~> 1.0.5"
   backend "s3" {
     key                   = "cloudwatch_logs.tfstate"
+    encrypt               = true
   }
 }