+++ terraform/environments/network/local.tf
@@ -1,3 +1,0 @@
-//locals {
-//  az_count = length(data.aws_availability_zones.available.names)
-//}