+++ CHANGELOG.md
@@ -1,11 +1,17 @@
 ## CHANGE LOG
 BonusBits Orchestrator
 
+## 2.3.2 - 09/20/2021 - Levon Becker
+* Added CloudFormation Template for creating tfstate bucket
+* Added/Updated Documentation for the Tfstate S3 Bucket setup
+
 ## 2.3.1 - 09/09/2021 - Levon Becker
 * Added aws secrets manager setup doc
 * Added aws secrets manager secret example json
 * Added aws secret permission policy example json
 * Updated ssh key create to not include username or hostname
+* Updated all roles to use encryption with the s3 backend tfstate bucket
+* Updated s3 download and upload tasks/libs to use s3 server side encryption
 
 ## 2.3.0 - 09/09/2021 - Levon Becker
 #### Minor Release: More Modules & Role Examples